Implemented an abstracted framework for the game.
@Rishabh, please implement three functions problem(), choice(), and message(), they are marked with a TODO, it shows up in spyder as an
exclamation mark(!) in the same spot as warnings and errors, I'll edit the wrappers to match your implementation after that.
Change the constants as required and copy paste all this code to the game file, we'll break it up into modules later.

Call me if you need help, I'll explain my design for the aforementioned three functions.